Rapid gun, borders, enemy hp, damage
Created the rapid gun with mouse aiming, tweakable fire rate and bullet speed.
Changed submarine movement and screen borders.
Added HP to the enemies and submarine, and damage to the projectiles. Also changed the destruction of the projectiles into the enemy class instead of their own, to avoid one extra collision function call.